fix(module8-task1): usePageInit update
