+++ project/src/hooks/usePageInit.ts
@@ -1,6 +1,8 @@
 import { useEffect } from 'react';
-import { useAppDispatch } from '.';
+import { useAppDispatch, useAppSelector } from '.';
 import { fetchCommentsAction, fetchHotelAction, fetchNearbyAction } from '../store/api-actions';
+import { authStatusSelector, citySelector, commentsSelector, nearbySelector, offerSelector } from '../store/selectors';
+import { Offer } from '../types';
 
 export const usePageInit = (id?: string) => {
   const dispatch = useAppDispatch();
@@ -11,4 +13,13 @@ export const usePageInit = (id?: string) => {
       dispatch(fetchNearbyAction(Number(id)));
     }
   }, [dispatch, id]);
+
+  const city = useAppSelector(citySelector);
+  const offer = useAppSelector(offerSelector);
+  const comments = useAppSelector(commentsSelector);
+  const authStatus = useAppSelector(authStatusSelector);
+  const nearby = useAppSelector(nearbySelector);
+  const points = nearby.map((item: Offer) => item.location);
+
+  return { city, offer, comments, authStatus, nearby, points };
 };
